pyproject.toml
[tool.poetry]
name = "django-project2"
version = "0.1.0"
description = ""
authors = ["akelaPro <ilya_chernov_2022@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
daphne = "^4.2.0"
psycopg2 = "^2.9.10"
python-decouple = "^3.8"
django-extensions = "^4.1"
django-imagekit = "^5.0.0"
dj-database-url = "^3.0.0"
djoser = "^2.3.1"
djangorestframework-simplejwt = "^5.5.0"
drf-nested-routers = "^0.94.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
